Create dualpal
This has quite a bit of extra stuff in it. 

I made my own convert from base to base function as well as helper functions for displaying information. 

The idea is that I wanted to use memoization to speed up the process of converting bases without using any built-in libraries etc. 

-Tyler