Create crypt1.java
In my opinion the best way to solve this one is to use permutations. 

Then and only then will you be able to test all the cases for the ENTIRE solution being a cryptarithm. 

In this scenario an ENTIRE Solution includes the partial products and the product itself. 

So even if the product is the same as another product...it may be a unique solution if you got that product via a unique couple of partial products. 

Permutation is the safest way not to miss anything. 

Take a look at my systemPermutation() function to understand the nested loops required to get all the work done. 

-Tyler